feat/user us3 trisno

See merge request apap2023/kelas-a-tk/ta_a_ayu_1!11

+++ order/src/main/java/com/apapedia/order/model/Cart.java
@@ -3,6 +3,7 @@ package com.apapedia.order.model;
 import java.util.List;
 import java.util.UUID;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import org.hibernate.annotations.GenericGenerator;
 
 import jakarta.persistence.CascadeType;
@@ -27,20 +28,19 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Cart {
-    
 
     @Id
     @GeneratedValue( generator = "cart_sequence")
     @GenericGenerator(name = "cart_sequence", strategy = "org.hibernate.id.UUIDGenerator")
     @Column(name = "id_cart", nullable = false)
 
-    private UUID idCart ;
+    private UUID idCart;
 
     @Column(name = "id_user", nullable = false)
     private UUID userId;
 
     @Column(name = "total_price", nullable = false)
-    private Integer totalPrice ; 
+    private long totalPrice;
 
     @OneToMany(mappedBy = "cart", fetch = FetchType.LAZY, cascade = CascadeType.ALL , orphanRemoval = true)
     private List<CartItem> listCartItem;
